feat: add setting response schema (draft)
